Rearranging some locations
